requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
future==0.17.1
importlib-metadata==0.19
jedi==0.14.1
kafka-python==1.4.7
more-itertools==7.2.0
parso==0.5.1
pluggy==0.12.0
pycparser==2.19
pymongo==3.8.0
python-jsonrpc-server==0.2.0
python-language-server==0.28.2
redis==3.2.1
six==1.12.0
tornado==6.0.3
zipp==0.6.0
